seekr/requirements.txt
rich
pandas
psutil
requests
scapy
speedtest-cli
matplotlib
GPUtil